backend/requirements.txt
google-generativeai
pydantic
openai
python-dotenv
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
asyncpg
pyjwt
psycopg2-binary
# psycopg2